Move and rename some files
